DOOM: Fixed MRGC-G32 resolution

+++ prboom-go/components/prboom/r_data.h
@@ -61,7 +61,7 @@ typedef struct
 typedef struct
 {
   char  name[8];         // Keep name for switch changing, etc.
-  int   next, index;     // killough 1/31/98: used in hashing algorithm
+  short next, index;     // killough 1/31/98: used in hashing algorithm
   // CPhipps - moved arrays with per-texture entries to elements here
   unsigned  widthmask;
   // CPhipps - end of additions
@@ -88,7 +88,7 @@ void R_PrecacheLevel (void);
 int R_FlatNumForName (const char* name);   // killough -- const added
 
 
-// R_*TextureNumForName returns the texture number for the texture name, or NO_TEXTURE if 
+// R_*TextureNumForName returns the texture number for the texture name, or NO_TEXTURE if
 //  there is no texture (i.e. "-") specified.
 /* cph 2006/07/23 - defined value for no-texture marker (texture "-" in the WAD file) */
 #define NO_TEXTURE 0
